pyproject.toml
[project]
name = "job_application_assistant"
version = "2.0.0"
description = "AI-powered job application assistant"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.201.1,<1.0.0",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.22",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
    "reportlab>=4.4.4",
    "PyPDF2>=3.0.1",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
    "pymilvus>=2.6.2",
    "sentence-transformers>=5.1.1",
    "requests>=2.32.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/job_application_assistant"]

[tool.crewai]
type = "crew"
